Crate libriakv

riakv

Log structured, append only, key value store implementation from Rust In Action with some enhancements.

Features

Persitent key value store with a hash table index

crc32 checksum validation for every key value pair stored.

Optionally, persitent index for fast loading

Exhaustive, comprehensive tests

Structs

KeyValuePair
Representation of a key value pair

RiaKV
Generic representation of a key value store in libriakv The underlying storage can be any type with the Read + Write + Seek trait bounds. The index attribute is used to maintain a mapping from the key value pairs to the position at which they are stored in the backing storage file.

Enums

IndexOp
Represent the kind of index operation to use for a given (KeyValuePair, u64) received during iterating over the contents of the storage file.

Type Definitions

ByteStr
Type to represent binary content internally

ByteString
Type to represent binary content